[BUGFIX] Re-rename module to `TYPO3/CMS/Screenshots/ScreenshotsManager`
Javascript modules loaded via requireJS should not use the PHP
namespace of the TYPO3 extension, i.e. not
'TYPO3/Documentation/Screenshots', but always the vendor and product
name of the TYPO3 CMS, so that it is derived from the following
pattern:

TYPO3/CMS/{extension}/{fileName}

This regression was introduced by commit #c2ca32a.